// forms-web/src/app/generated/model/question-dependency.ts
/**
 * fbuilder APIs
 *
 * Contact: admin@fbuilder.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import {FormArray, FormControl, FormGroup, Validators} from "@angular/forms";
import {DocumentData, QueryDocumentSnapshot, SnapshotOptions, Timestamp} from "@angular/fire/firestore";

export interface QuestionDependency {
        /**
        * The id of question
        */
        questionId: string;

        answer: string;

}



// Firestore data converter for QuestionDependency
export const QuestionDependencyFirestoreConverter = {
toFirestore(modelObject: QuestionDependency): DocumentData {
const result: DocumentData = {};
        if (modelObject.questionId !== undefined) {
                    result['questionId'] = modelObject.questionId;
        }
        if (modelObject.answer !== undefined) {
                    result['answer'] = modelObject.answer;
        }
return result;
},
fromFirestore(snapshot: QueryDocumentSnapshot, options: SnapshotOptions): QuestionDependency {
const data = snapshot.data(options);
return {

                    questionId: data['questionId'],

                    answer: data['answer'],
    } as QuestionDependency;
    }

};