// forms-web/src/app/generated/model/question-option.ts
/**
 * fbuilder APIs
 *
 * Contact: admin@fbuilder.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import {FormArray, FormControl, FormGroup, Validators} from "@angular/forms";
import {DocumentData, QueryDocumentSnapshot, SnapshotOptions, Timestamp} from "@angular/fire/firestore";

export interface QuestionOption {
        optionId?: string;

        /**
        * The option of a question
        */
        text: string;

        nextQuestionIds?: Array<string>;

}

export type QuestionOptionFormType = FormGroup<{
    text: FormControl<string>;
 }>

export function getQuestionOptionForm(): QuestionOptionFormType {
    return new FormGroup({

    text: new FormControl<string>("", {  nonNullable:  true  ,
    validators: [  Validators.required,  Validators.minLength(1),] } ),


    })
}


// Firestore data converter for QuestionOption
export const QuestionOptionFirestoreConverter = {
toFirestore(modelObject: QuestionOption): DocumentData {
const result: DocumentData = {};
        if (modelObject.optionId !== undefined) {
                    result['optionId'] = modelObject.optionId;
        }
        if (modelObject.text !== undefined) {
                    result['text'] = modelObject.text;
        }
        if (modelObject.nextQuestionIds !== undefined) {
                    result['nextQuestionIds'] = modelObject.nextQuestionIds;
        }
return result;
},
fromFirestore(snapshot: QueryDocumentSnapshot, options: SnapshotOptions): QuestionOption {
const data = snapshot.data(options);
return {

                    optionId: data['optionId'],

                    text: data['text'],

                    nextQuestionIds: data['nextQuestionIds'],
    } as QuestionOption;
    }

};