// forms-web/src/app/generated/model/question-type.ts
/**
 * fbuilder APIs
 *
 * Contact: admin@fbuilder.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Question Type.
 */
export type QuestionType = 'Single Choice' | 'Multiple Choice' | 'Open Ended' | 'Yes/No';

export const QuestionType = {

    SingleChoice: 'Single Choice' as QuestionType,

    MultipleChoice: 'Multiple Choice' as QuestionType,

    OpenEnded: 'Open Ended' as QuestionType,

    YesNo: 'Yes/No' as QuestionType
};